--- a/.gitmodules
+++ b/.gitmodules
@@ -4,3 +4,6 @@
 [submodule "numpy/core/src/umath/svml"]
 	path = numpy/core/src/umath/svml
 	url = https://github.com/numpy/SVML.git
+[submodule "numpy/core/src/umath/optimized-routines"]
+	path = numpy/core/src/umath/optimized-routines
+	url = https://github.com/ARM-software/optimized-routines.git