--- a/numpy/core/src/umath/loops_umath_fp.dispatch.c.src
+++ b/numpy/core/src/umath/loops_umath_fp.dispatch.c.src
@@ -1,6 +1,12 @@
 /*@targets
- ** $maxopt baseline avx512_skx
+ ** $maxopt baseline
+ ** avx512_skx
+ ** asimd
  */
+/*
+ * Copyright 2023 Arm Limited and/or its affiliates <open-source-office@arm.com>
+ */
+
 #include "numpy/npy_math.h"
 #include "simd/simd.h"
 #include "loops_utils.h"
@@ -8,7 +14,17 @@
 #include "npy_svml.h"
 #include "fast_loop_macros.h"
 
-#if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#if NPY_SIMD
+#if defined(NPY_HAVE_ASIMD) && defined(NPY_CAN_LINK_AOR)
+#include <mathlib.h>
+#define _NPY_OPTIMIZED_ROUTINES
+#endif
+#if defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#define _NPY_SVML
+#endif
+#endif
+
+#if defined(_NPY_SVML)
 /**begin repeat
  * #sfx = f32, f64#
  * #func_suffix = f16, 8#
@@ -58,7 +74,9 @@ simd_@func@_@sfx@(const npyv_lanetype_@sfx@ *src, npy_intp ssrc,
 }
 /**end repeat1**/
 /**end repeat**/
+#endif
 
+#if defined(_NPY_OPTIMIZED_ROUTINES) || defined(_NPY_SVML)
 /**begin repeat
  * #func = sin, cos#
  */
@@ -74,7 +92,11 @@ simd_@func@_f64(const double *src, npy_intp ssrc,
         } else {
             x = npyv_loadn_tillz_f64(src, ssrc, len);
         }
+        #if defined(_NPY_OPTIMIZED_ROUTINES)
+        npyv_f64 out = __v_@func@(x);
+        #else
         npyv_f64 out = __svml_@func@8(x);
+        #endif
         if (sdst == 1) {
             npyv_store_till_f64(dst, len, out);
         } else {
@@ -84,7 +106,9 @@ simd_@func@_f64(const double *src, npy_intp ssrc,
     npyv_cleanup();
 }
 /**end repeat**/
+#endif
 
+#if defined(_NPY_SVML)
 /**begin repeat
  * #sfx = f32, f64#
  * #func_suffix = f16, 8#
@@ -92,7 +116,6 @@ simd_@func@_f64(const double *src, npy_intp ssrc,
 /**begin repeat1
  * #func = pow, atan2#
  */
-
 static void
 simd_@func@_@sfx@(const npyv_lanetype_@sfx@ *src1, npy_intp ssrc1,
                   const npyv_lanetype_@sfx@ *src2, npy_intp ssrc2,
@@ -182,7 +205,7 @@ avx512_@func@_f16(const npy_half *src, npy_half *dst, npy_intp len)
 NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(HALF_@func@)
 (char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(data))
 {
-#if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#if defined(_NPY_SVML)
     const npy_half *src = (npy_half*)args[0];
           npy_half *dst = (npy_half*)args[1];
     const int lsize = sizeof(src[0]);
@@ -216,7 +239,7 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(HALF_@func@)
 NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
 (char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(data))
 {
-#if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#if defined(_NPY_SVML)
     const @type@ *src = (@type@*)args[0];
           @type@ *dst = (@type@*)args[1];
     const int lsize = sizeof(src[0]);
@@ -252,7 +275,7 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
 NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
 (char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(data))
 {
-#if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#if defined(_NPY_SVML)
     const @type@ *src1 = (@type@*)args[0];
     const @type@ *src2 = (@type@*)args[1];
           @type@ *dst  = (@type@*)args[2];
@@ -284,7 +307,7 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
 NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(DOUBLE_@func@)
 (char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(data))
 {
-#if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
+#if defined(_NPY_SVML) || defined(_NPY_OPTIMIZED_ROUTINES)
     const double *src = (double*)args[0];
           double *dst = (double*)args[1];
     const int lsize = sizeof(src[0]);