--- a/tools/ci/cirrus_general.yml
+++ b/tools/ci/cirrus_general.yml
@@ -1,3 +1,5 @@
+# Copyright 2023 Arm Limited and/or its affiliates <open-source-office@arm.com>
+
 build_and_store_wheels: &BUILD_AND_STORE_WHEELS
   install_cibuildwheel_script:
     - python -m pip install cibuildwheel==2.12.0
@@ -34,6 +36,7 @@ linux_aarch64_task:
   build_script: |
     apt install -y python3-venv python-is-python3 gfortran libatlas-base-dev libgfortran5 eatmydata
     git fetch origin
+    git submodule update --init
     ./tools/travis-before-install.sh
     which python
     echo $CIRRUS_CHANGE_MESSAGE