@Mousius 2023-02-08
> Slightly apologetic comment - AOR/pl has recently removed some symbols, and likely AOR/math will at some point soon as well, so the way function names are chosen is not future-proof (and won't work at all if you want any routines from pl/). How are you building this? I don't see any config.mk file in this patch, so interested in how AOR is being configured?

I'm just building the bits I need as a submodule in `numpy` rather than taking the whole library in, that would require additional effort that doesn't appear to be justified?