@Mousius 2023-02-08
> > > Slightly apologetic comment - AOR/pl has recently removed some symbols, and likely AOR/math will at some point soon as well, so the way function names are chosen is not future-proof (and won't work at all if you want any routines from pl/). How are you building this? I don't see any config.mk file in this patch, so interested in how AOR is being configured?
> > 
> > 
> > I'm just building the bits I need as a submodule in `numpy` rather than taking the whole library in, that would require additional effort that doesn't appear to be justified?
> 
> So the C files are just compiled independently of the AOR Makefiles? Makes sense, but since AOR is designed to be build with a config file, you may end up missing certain features (or even routines) that are enabled based on settings in the config file, for instance without `-DWANT_SIMD_EXCEPT=1` you will not have fp exceptions triggered correctly. My understanding is that numpy requires this, so I'm surprised not to see it mentioned here

Likewise, but if the tests are passing I would consider that sufficient for now, I'm keen not to add too many additional flags if I can avoid it especially given they're not namespaced to Optimized Routines so could cause other software to compile weirdly.