@Mousius 2023-02-08
> > > > > Slightly apologetic comment - AOR/pl has recently removed some symbols, and likely AOR/math will at some point soon as well, so the way function names are chosen is not future-proof (and won't work at all if you want any routines from pl/). How are you building this? I don't see any config.mk file in this patch, so interested in how AOR is being configured?
> > > > 
> > > > 
> > > > I'm just building the bits I need as a submodule in `numpy` rather than taking the whole library in, that would require additional effort that doesn't appear to be justified?
> > > 
> > > 
> > > So the C files are just compiled independently of the AOR Makefiles? Makes sense, but since AOR is designed to be build with a config file, you may end up missing certain features (or even routines) that are enabled based on settings in the config file, for instance without `-DWANT_SIMD_EXCEPT=1` you will not have fp exceptions triggered correctly. My understanding is that numpy requires this, so I'm surprised not to see it mentioned here
> > 
> > 
> > Likewise, but if the tests are passing I would consider that sufficient for now, I'm keen not to add too many additional flags if I can avoid it especially given they're not namespaced to Optimized Routines so could cause other software to compile weirdly.
> 
> I understand what you're saying, but I think this could become very difficult to maintain for a couple of reasons. One is that AOR routines are not always completely self-contained in one source file - there may be helper routines or coefficient arrays in a different TU (not the case for sin/cos, so it was fine for this patch). These get moved around from time to time - we don't make any promises about what files will exist, and indeed we are in the process of shuffling some things around. The VFABI-mangled symbols don't come from `v_<func>.c` at all, they come from `vn_<func>.c` but again, we are in the process of changing this. The only way to reliably get the symbols you need is by building the whole library.
> 
> As well, there are certain flags that at some point you will need in your compiler invocation, for instance you can't build SVE routines without explicitly enabling them, either in the config file or with `-DWANT_SVE_MATH=1`.

I think that's ok as from the brief discussion with @mattip, the focus should be on moving to universal intrinsics in `numpy`. AOR is a good intermediate step for the performance boost but we can avoid long-term dependence on it if we can't minimally include it. In the intermediary state we can accept having the files as-is and a slightly painful upgrade path, that'll likely land on me 😸